+++ R/predictions.R
@@ -1,6 +1,6 @@
 # select prediction method, based on model-object
 #' @importFrom sjstats link_inverse
-select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, faminfo, ppd, terms, typical, ...) {
+select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, faminfo, ppd, terms, typical, prettify, prettify.at, ...) {
   # get link-inverse-function
   linv <- sjstats::link_inverse(model)
 
@@ -12,10 +12,10 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
     fitfram <- get_predictions_svyglmnb(model, expanded_frame, ci.lvl, linv, ...)
   } else if (fun == "stanreg") {
     # stan-objects -----
-    fitfram <- get_predictions_stan(model, expanded_frame, ci.lvl, type, faminfo, ppd, ...)
+    fitfram <- get_predictions_stan(model, expanded_frame, ci.lvl, type, faminfo, ppd, terms, ...)
   } else if (fun == "brmsfit") {
     # brms-objects -----
-    fitfram <- get_predictions_stan(model, expanded_frame, ci.lvl, type, faminfo, ppd, ...)
+    fitfram <- get_predictions_stan(model, expanded_frame, ci.lvl, type, faminfo, ppd, terms, ...)
   } else if (fun == "coxph") {
     # coxph-objects -----
     fitfram <- get_predictions_coxph(model, expanded_frame, ci.lvl, ...)
@@ -27,7 +27,7 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
     fitfram <- get_predictions_glmmTMB(model, expanded_frame, ci.lvl, linv, ...)
   } else if (fun %in% c("lmer", "nlmer", "glmer")) {
     # merMod-objects  -----
-    fitfram <- get_predictions_merMod(model, expanded_frame, ci.lvl, linv, type, terms, typical, ...)
+    fitfram <- get_predictions_merMod(model, expanded_frame, ci.lvl, linv, type, terms, typical, prettify, prettify.at, ...)
   } else if (fun == "gam") {
     # gam-objects -----
     fitfram <- get_predictions_gam(model, expanded_frame, ci.lvl, linv, ...)
@@ -36,7 +36,7 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
     fitfram <- get_predictions_vgam(model, expanded_frame, ci.lvl, linv, ...)
   } else if (fun %in% c("lme", "gls", "plm")) {
     # lme-objects -----
-    fitfram <- get_predictions_lme(model, expanded_frame, ci.lvl, terms, typical, ...)
+    fitfram <- get_predictions_lme(model, expanded_frame, ci.lvl, terms, typical, prettify, prettify.at, ...)
   } else if (fun == "gee") {
     # gee-objects -----
     fitfram <- get_predictions_gee(model, expanded_frame, linv, ...)
@@ -54,7 +54,7 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
     fitfram <- get_predictions_polr(model, expanded_frame, linv, ...)
   } else if (fun %in% c("betareg", "truncreg", "zeroinfl", "hurdle")) {
     # betareg, truncreg, zeroinfl and hurdle-objects -----
-    fitfram <- get_predictions_generic2(model, expanded_frame, ci.lvl, fun, typical, terms, ...)
+    fitfram <- get_predictions_generic2(model, expanded_frame, ci.lvl, fun, typical, terms, prettify, prettify.at, ...)
   } else if (fun %in% c("glm", "glm.nb")) {
     # glm-objects -----
     fitfram <- get_predictions_glm(model, expanded_frame, ci.lvl, linv, ...)
@@ -289,7 +289,7 @@ get_predictions_clm <- function(model, fitfram, ci.lvl, linv, ...) {
 
 # predictions for regression models w/o SE ----
 
-get_predictions_generic2 <- function(model, fitfram, ci.lvl, fun, typical, terms, ...) {
+get_predictions_generic2 <- function(model, fitfram, ci.lvl, fun, typical, terms, prettify, prettify.at, ...) {
   # get prediction type.
   pt <- dplyr::case_when(
     fun %in% c("hurdle", "zeroinfl") ~ "response",
@@ -320,7 +320,9 @@ get_predictions_generic2 <- function(model, fitfram, ci.lvl, fun, typical, terms
       fitfram = fitfram,
       typical = typical,
       terms = terms,
-      fun = fun
+      fun = fun,
+      prettify = prettify,
+      prettify.at = prettify.at
     )
 
   se.fit <- se.pred$se.fit
@@ -441,7 +443,7 @@ get_predictions_glmmTMB <- function(model, fitfram, ci.lvl, linv, ...) {
 
 # predictions for merMod ----
 
-get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, typical, ...) {
+get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, typical, prettify, prettify.at, ...) {
   # does user want standard errors?
   se <- !is.null(ci.lvl) && !is.na(ci.lvl)
 
@@ -475,7 +477,9 @@ get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, ty
         fitfram = fitfram,
         typical = typical,
         terms = terms,
-        type = type
+        type = type,
+        prettify = prettify,
+        prettify.at = prettify.at
       )
 
     se.fit <- se.pred$se.fit
@@ -510,13 +514,15 @@ get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, ty
 
 # predictions for stan ----
 
+#' @importFrom tidyr gather
 #' @importFrom tibble as_tibble
-#' @importFrom sjstats hdi resp_var
+#' @importFrom sjstats hdi resp_var resp_val
 #' @importFrom sjmisc rotate_df
 #' @importFrom purrr map_dbl map_df
-#' @importFrom dplyr bind_cols
-#' @importFrom stats median
-get_predictions_stan <- function(model, fitfram, ci.lvl, type, faminfo, ppd, ...) {
+#' @importFrom dplyr bind_cols select bind_rows n_distinct
+#' @importFrom stats median formula
+#' @importFrom tidyselect ends_with
+get_predictions_stan <- function(model, fitfram, ci.lvl, type, faminfo, ppd, terms, ...) {
   # check if pkg is available
   if (!requireNamespace("rstantools", quietly = TRUE)) {
     stop("Package `rstantools` is required to compute predictions.", call. = F)
@@ -579,8 +585,47 @@ get_predictions_stan <- function(model, fitfram, ci.lvl, type, faminfo, ppd, ...
   # we have a list of 4000 samples, so we need to coerce to data frame
   prdat <- tibble::as_tibble(prdat)
 
-  # compute median, as "most probable estimate"
-  fitfram$predicted <- purrr::map_dbl(prdat, stats::median)
+
+  # handle cumulative link models
+
+  if (inherits(model, "brmsfit") && faminfo$family == "cumulative") {
+
+    tmp <- prdat %>%
+      purrr::map_df(stats::median) %>%
+      tidyr::gather(key = "grp", value = "predicted")
+
+    resp.vals <- sort(sjstats::resp_val(model))
+    term.cats <- nrow(fitfram)
+    resp.cats <- dplyr::n_distinct(resp.vals, na.rm = TRUE)
+    fitfram <- purrr::map_df(1:resp.cats, ~ fitfram)
+
+    fitfram$response.level <- rep(unique(resp.vals), each = term.cats)
+    fitfram$predicted <- tmp$predicted
+
+  } else if (inherits(model, "brmsfit") && !is.null(stats::formula(model)$responses)) {
+
+    # handle multivariate response models
+
+    tmp <- prdat %>%
+      purrr::map_df(stats::median) %>%
+      tidyr::gather(key = "grp", value = "predicted")
+
+    resp.vars <- sjstats::resp_var(model)
+    fitfram <- purrr::map_df(1:length(resp.vars), ~ fitfram)
+    fitfram$response.level <- ""
+
+    for (i in resp.vars) {
+      pos <- tidyselect::ends_with(i, vars = tmp$grp)
+      fitfram$response.level[pos] <- i
+    }
+
+    fitfram$predicted <- tmp$predicted
+
+  } else {
+    # compute median, as "most probable estimate"
+    fitfram$predicted <- purrr::map_dbl(prdat, stats::median)
+  }
+
 
   # for posterior predictive distributions, we compute
   # the predictive intervals
@@ -717,6 +762,8 @@ get_predictions_vgam <- function(model, fitfram, ci.lvl, linv, ...) {
 
 # predictions for lm ----
 
+#' @importFrom dplyr bind_cols
+#' @importFrom tidyr gather
 get_predictions_lm <- function(model, fitfram, ci.lvl, linv, ...) {
   # does user want standard errors?
   se <- !is.null(ci.lvl) && !is.na(ci.lvl)
@@ -745,8 +792,22 @@ get_predictions_lm <- function(model, fitfram, ci.lvl, linv, ...) {
     fitfram$conf.low <- prdat$fit - stats::qnorm(ci) * prdat$se.fit
     fitfram$conf.high <- prdat$fit + stats::qnorm(ci) * prdat$se.fit
   } else {
-    # copy predictions
-    fitfram$predicted <- as.vector(prdat)
+    # check if we have a multivariate response model
+    pdim <- dim(prdat)
+    if (!is.null(pdim) && pdim[2] > 1) {
+      tmp <- dplyr::bind_cols(fitfram, as.data.frame(prdat))
+      gather.vars <- (ncol(fitfram) + 1):ncol(tmp)
+
+      fitfram <- tidyr::gather(
+        tmp,
+        key = "response.level",
+        value = "predicted",
+        !! gather.vars
+      )
+    } else {
+      # copy predictions
+      fitfram$predicted <- as.vector(prdat)
+    }
 
     # no CI
     fitfram$conf.low <- NA
@@ -763,7 +824,7 @@ get_predictions_lm <- function(model, fitfram, ci.lvl, linv, ...) {
 #' @importFrom sjstats resp_var pred_vars
 #' @importFrom purrr map
 #' @importFrom tibble add_column
-get_predictions_lme <- function(model, fitfram, ci.lvl, terms, typical, ...) {
+get_predictions_lme <- function(model, fitfram, ci.lvl, terms, typical, prettify, prettify.at, ...) {
   # does user want standard errors?
   se <- !is.null(ci.lvl) && !is.na(ci.lvl)
 
@@ -787,7 +848,15 @@ get_predictions_lme <- function(model, fitfram, ci.lvl, terms, typical, ...) {
 
   # did user request standard errors? if yes, compute CI
   if (se) {
-    se.pred <- get_se_from_vcov(model = model, fitfram = fitfram, typical = typical, terms = terms)
+    se.pred <-
+      get_se_from_vcov(
+        model = model,
+        fitfram = fitfram,
+        typical = typical,
+        terms = terms,
+        prettify = prettify,
+        prettify.at = prettify.at
+      )
 
     se.fit <- se.pred$se.fit
     fitfram <- se.pred$fitfram
@@ -918,7 +987,7 @@ get_base_fitfram <- function(fitfram, linv, prdat, se, ci.lvl) {
 #' @importFrom dplyr arrange
 #' @importFrom sjstats resp_var model_frame
 #' @importFrom rlang parse_expr
-get_se_from_vcov <- function(model, fitfram, typical, terms, fun = NULL, type = "fe") {
+get_se_from_vcov <- function(model, fitfram, typical, terms, fun = NULL, type = "fe", prettify = TRUE, prettify.at = 25) {
   # copy data frame with predictions
   newdata <- get_expanded_data(
     model,
@@ -926,7 +995,9 @@ get_se_from_vcov <- function(model, fitfram, typical, terms, fun = NULL, type = 
     terms,
     typ.fun = typical,
     fac.typical = FALSE,
-    type = type
+    type = type,
+    prettify = prettify,
+    prettify.at = prettify.at
   )
 
   # add response